wip: merge http.ex and server.ex
