+++ lib/tidewave/mcp/server.ex
@@ -3,6 +3,7 @@ defmodule Tidewave.MCP.Server do
 
   require Logger
 
+  import Plug.Conn
   alias Tidewave.MCP.Tools
 
   @protocol_version "2025-03-06"
@@ -197,7 +198,121 @@ defmodule Tidewave.MCP.Server do
     end
   end
 
-  ## handle_message function for SSE plug
+  ## HTTP transport functions
+
+  def handle_http_message(conn) do
+    Logger.info("Received #{conn.method} message")
+    params = conn.body_params
+    conn = fetch_query_params(conn)
+    Logger.debug("Raw params: #{inspect(params, pretty: true)}")
+
+    case validate_jsonrpc_message(params) do
+      {:ok, message} ->
+        handle_http_jsonrpc_message(conn, message)
+
+      {:error, :invalid_jsonrpc} ->
+        Logger.warning("Invalid JSON-RPC message format")
+        send_jsonrpc_error(conn, nil, -32600, "Could not parse message")
+    end
+  end
+
+  defp handle_http_jsonrpc_message(conn, message) do
+    assigns = %{connect_params: conn.query_params}
+    assigns = Map.merge(assigns, conn.private.tidewave_config)
+
+    case message do
+      # Handle initialization sequence
+      %{"method" => "initialize"} = msg ->
+        Logger.info("Routing MCP message - Method: initialize, ID: #{msg["id"]}")
+        Logger.debug("Full message: #{inspect(msg, pretty: true)}")
+        {:ok, response} = handle_message(msg, assigns)
+
+        conn |> put_status(200) |> send_json(response)
+
+      %{"method" => "notifications/initialized"} ->
+        conn |> put_status(202) |> send_json(%{status: "ok"})
+
+      %{"method" => "notifications/cancelled"} ->
+        # Just log the cancellation notification and return ok
+        Logger.info("Request cancelled: #{inspect(message["params"])}")
+        conn |> put_status(202) |> send_json(%{status: "ok"})
+
+      # Handle requests
+      %{"id" => _} ->
+        handle_http_request(conn, message, assigns)
+
+      # For any other notifications without id
+      _ ->
+        conn |> put_status(202) |> send_json(%{status: "ok"})
+    end
+  end
+
+  defp handle_http_request(conn, message, assigns) do
+    case handle_message(message, assigns) do
+      {:ok, nil} ->
+        conn |> put_status(202) |> send_json(%{status: "ok"})
+
+      {:ok, response} ->
+        Logger.debug("Sending HTTP response: #{inspect(response, pretty: true)}")
+        conn |> put_status(200) |> send_json(response)
+
+      {:error, error_response} ->
+        Logger.warning("Error handling message: #{inspect(error_response)}")
+        conn |> put_status(400) |> send_json(error_response)
+    end
+  end
+
+  defp send_json(conn, data) do
+    conn
+    |> put_resp_content_type("application/json")
+    |> send_resp(conn.status || 200, Jason.encode!(data))
+  end
+
+  defp send_jsonrpc_error(conn, id, code, message, data \\ nil) do
+    error = %{
+      code: code,
+      message: message
+    }
+
+    error = if data, do: Map.put(error, :data, data), else: error
+
+    response = %{
+      jsonrpc: "2.0",
+      id: id,
+      error: error
+    }
+
+    conn
+    |> put_resp_content_type("application/json")
+    |> send_resp(200, Jason.encode!(response))
+  end
+
+  defp validate_jsonrpc_message(%{"jsonrpc" => "2.0"} = message) do
+    cond do
+      # Request must have method and id (string or number)
+      Map.has_key?(message, "id") and Map.has_key?(message, "method") ->
+        case message["id"] do
+          id when is_binary(id) or is_number(id) -> {:ok, message}
+          nil -> {:error, :invalid_jsonrpc}
+          _ -> {:error, :invalid_jsonrpc}
+        end
+
+      # Notification must have method but no id
+      not Map.has_key?(message, "id") and Map.has_key?(message, "method") ->
+        {:ok, message}
+
+      # reply (e.g. to ping) with ID + result
+      Map.has_key?(message, "id") and Map.has_key?(message, "result") ->
+        {:ok, message}
+
+      true ->
+        {:error, :invalid_jsonrpc}
+    end
+  end
+
+  defp validate_jsonrpc_message(_), do: {:error, :invalid_jsonrpc}
+
+  ## handle_message function for internal MCP message processing
 
   # Built-in message routing
   def handle_message(%{"method" => "notifications/initialized"} = message, _assigns) do
